// flutter/js/src/proto/message.ts
export enum ControlKey {
  Unknown = 0,
  Alt = 1,
  Backspace = 2,
  CapsLock = 3,
  Control = 4,
  Delete = 5,
  DownArrow = 6,
  End = 7,
  Escape = 8,
  F1 = 9,
  F10 = 10,
  F11 = 11,
  F12 = 12,
  F2 = 13,
  F3 = 14,
  F4 = 15,
  F5 = 16,
  F6 = 17,
  F7 = 18,
  F8 = 19,
  F9 = 20,
  Home = 21,
  LeftArrow = 22,
  /** Meta - / meta key (also known as "windows"; "super"; and "command") */
  Meta = 23,
  /** Option - / option key on macOS (alt key on Linux and Windows) */
  Option = 24,
  PageDown = 25,
  PageUp = 26,
  Return = 27,
  RightArrow = 28,
  Shift = 29,
  Space = 30,
  Tab = 31,
  UpArrow = 32,
  Numpad0 = 33,
  Numpad1 = 34,
  Numpad2 = 35,
  Numpad3 = 36,
  Numpad4 = 37,
  Numpad5 = 38,
  Numpad6 = 39,
  Numpad7 = 40,
  Numpad8 = 41,
  Numpad9 = 42,
  Cancel = 43,
  Clear = 44,
  /** Menu - deprecated, use Alt instead */
  Menu = 45,
  Pause = 46,
  Kana = 47,
  Hangul = 48,
  Junja = 49,
  Final = 50,
  Hanja = 51,
  Kanji = 52,
  Convert = 53,
  Select = 54,
  Print = 55,
  Execute = 56,
  Snapshot = 57,
  Insert = 58,
  Help = 59,
  Sleep = 60,
  Separator = 61,
  Scroll = 62,
  NumLock = 63,
  RWin = 64,
  Apps = 65,
  Multiply = 66,
  Add = 67,
  Subtract = 68,
  Decimal = 69,
  Divide = 70,
  Equals = 71,
  NumpadEnter = 72,
  RShift = 73,
  RControl = 74,
  RAlt = 75,
  /** VolumeMute - mainly used on mobile devices as controlled side */
  VolumeMute = 76,
  VolumeUp = 77,
  VolumeDown = 78,
  /** Power - mainly used on mobile devices as controlled side */
  Power = 79,
  CtrlAltDel = 100,
  LockScreen = 101,
  UNRECOGNIZED = -1,
}
// ...
export interface VideoFrame {
  vp9s?: EncodedVideoFrames | undefined;
  rgb?: RGB | undefined;
  yuv?: YUV | undefined;
  h264s?: EncodedVideoFrames | undefined;
  h265s?: EncodedVideoFrames | undefined;
  vp8s?: EncodedVideoFrames | undefined;
  av1s?: EncodedVideoFrames | undefined;
  display: number;
}

export interface IdPk {
  id: string;
  pk: Uint8Array;
}

export interface DisplayInfo {
  x: number;
  y: number;
  width: number;
  height: number;
  name: string;
  online: boolean;
  cursor_embedded: boolean;
  original_resolution: Resolution | undefined;
  scale: number;
}

export interface PortForward {
  host: string;
  port: number;
}

export interface FileTransfer {
  dir: string;
  show_hidden: boolean;
}

export interface OSLogin {
  username: string;
  password: string;
}

export interface LoginRequest {
  username: string;
  password: Uint8Array;
  my_id: string;
  my_name: string;
  option: OptionMessage | undefined;
  file_transfer?: FileTransfer | undefined;
  port_forward?: PortForward | undefined;
  video_ack_required: boolean;
  session_id: number;
  version: string;
  os_login: OSLogin | undefined;
  my_platform: string;
  hwid: Uint8Array;
}
// ...
export interface PeerInfo {
  username: string;
  hostname: string;
  platform: string;
  displays: DisplayInfo[];
  current_display: number;
  sas_enabled: boolean;
  version: string;
  features: Features | undefined;
  encoding: SupportedEncoding | undefined;
  resolutions:
    | SupportedResolutions
    | undefined;
  /**
   * Use JSON's key-value format which is friendly for peer to handle.
   * NOTE: Only support one-level dictionaries (for peer to update), and the key is of type string.
   */
  platform_additions: string;
  windows_sessions: WindowsSessions | undefined;
}

export interface WindowsSession {
  sid: number;
  name: string;
}

export interface LoginResponse {
  error?: string | undefined;
  peer_info?: PeerInfo | undefined;
  enable_trusted_devices: boolean;
}
// ...
export interface MouseEvent {
  mask: number;
  x: number;
  y: number;
  modifiers: ControlKey[];
}

export interface KeyEvent {
  /** `down` indicates the key's state(down or up). */
  down: boolean;
  /** `press` indicates a click event(down and up). */
  press: boolean;
  control_key?:
    | ControlKey
    | undefined;
  /** position key code. win: scancode, linux: key code, macos: key code */
  chr?: number | undefined;
  unicode?: number | undefined;
  seq?:
    | string
    | undefined;
  /**
   * high word. virtual keycode
   * low word. unicode
   */
  win2win_hotkey?: number | undefined;
  modifiers: ControlKey[];
  mode: KeyboardMode;
}

export interface CursorData {
  id: number;
  hotx: number;
  hoty: number;
  width: number;
  height: number;
  colors: Uint8Array;
}

export interface CursorPosition {
  x: number;
  y: number;
}

export interface Hash {
  salt: string;
  challenge: string;
}

export interface Clipboard {
  compress: boolean;
  content: Uint8Array;
  width: number;
  height: number;
  format: ClipboardFormat;
  /** Special format name, only used when format is Special. */
  special_name: string;
}
// ...
export interface SwitchDisplay {
  display: number;
  x: number;
  y: number;
  width: number;
  height: number;
  cursor_embedded: boolean;
  resolutions:
    | SupportedResolutions
    | undefined;
  /** Do not care about the origin point for now. */
  original_resolution: Resolution | undefined;
}

export interface CaptureDisplays {
  add: number[];
  sub: number[];
  set: number[];
}

export interface ToggleVirtualDisplay {
  display: number;
  on: boolean;
}

export interface TogglePrivacyMode {
  impl_key: string;
  on: boolean;
}

export interface PermissionInfo {
  permission: PermissionInfo_Permission;
  enabled: boolean;
}

export enum PermissionInfo_Permission {
  Keyboard = 0,
  Clipboard = 2,
  Audio = 3,
  File = 4,
  Restart = 5,
  Recording = 6,
  BlockInput = 7,
  UNRECOGNIZED = -1,
}
// ...
export interface OptionMessage {
  image_quality: ImageQuality;
  lock_after_session_end: OptionMessage_BoolOption;
  show_remote_cursor: OptionMessage_BoolOption;
  privacy_mode: OptionMessage_BoolOption;
  block_input: OptionMessage_BoolOption;
  custom_image_quality: number;
  disable_audio: OptionMessage_BoolOption;
  disable_clipboard: OptionMessage_BoolOption;
  enable_file_transfer: OptionMessage_BoolOption;
  supported_decoding: SupportedDecoding | undefined;
  custom_fps: number;
  disable_keyboard: OptionMessage_BoolOption;
  /**
   * Position 13 is used for Resolution. Remove later.
   * Resolution custom_resolution = 13;
   * BoolOption support_windows_specific_session = 14;
   * starting from 15 please, do not use removed fields
   */
  follow_remote_cursor: OptionMessage_BoolOption;
  follow_remote_window: OptionMessage_BoolOption;
}
// ...
export interface TestDelay {
  time: number;
  from_client: boolean;
  last_delay: number;
  target_bitrate: number;
}

export interface PublicKey {
  asymmetric_value: Uint8Array;
  symmetric_value: Uint8Array;
}

export interface SignedId {
  id: Uint8Array;
}

export interface AudioFormat {
  sample_rate: number;
  channels: number;
}

export interface AudioFrame {
  data: Uint8Array;
}
// ...
export interface SwitchSidesResponse {
  uuid: Uint8Array;
  lr: LoginRequest | undefined;
}

export interface SwitchBack {
}
// ...
export interface Misc {
  chat_message?: ChatMessage | undefined;
  switch_display?: SwitchDisplay | undefined;
  permission_info?: PermissionInfo | undefined;
  option?: OptionMessage | undefined;
  audio_format?: AudioFormat | undefined;
  close_reason?: string | undefined;
  refresh_video?: boolean | undefined;
  video_received?: boolean | undefined;
  back_notification?: BackNotification | undefined;
  restart_remote_device?: boolean | undefined;
  uac?: boolean | undefined;
  foreground_window_elevated?: boolean | undefined;
  stop_service?: boolean | undefined;
  elevation_request?: ElevationRequest | undefined;
  elevation_response?: string | undefined;
  portable_service_running?: boolean | undefined;
  switch_sides_request?: SwitchSidesRequest | undefined;
  switch_back?:
    | SwitchBack
    | undefined;
  /**
   * Deprecated since 1.2.4, use `change_display_resolution` (36) instead.
   * But we must keep it for compatibility when peer version < 1.2.4.
   */
  change_resolution?: Resolution | undefined;
  plugin_request?: PluginRequest | undefined;
  plugin_failure?:
    | PluginFailure
    | undefined;
  /** deprecated */
  full_speed_fps?: number | undefined;
  auto_adjust_fps?: number | undefined;
  client_record_status?: boolean | undefined;
  capture_displays?: CaptureDisplays | undefined;
  refresh_video_display?: number | undefined;
  toggle_virtual_display?: ToggleVirtualDisplay | undefined;
  toggle_privacy_mode?: TogglePrivacyMode | undefined;
  supported_encoding?: SupportedEncoding | undefined;
  selected_sid?: number | undefined;
  change_display_resolution?: DisplayResolution | undefined;
  message_query?: MessageQuery | undefined;
  follow_current_display?: number | undefined;
}

export interface VoiceCallRequest {
  req_timestamp: number;
  /** Indicates whether the request is a connect action or a disconnect action. */
  is_connect: boolean;
}

export interface VoiceCallResponse {
  accepted: boolean;
  /** Should copy from [VoiceCallRequest::req_timestamp]. */
  req_timestamp: number;
  ack_timestamp: number;
}

export interface Message {
  signed_id?: SignedId | undefined;
  public_key?: PublicKey | undefined;
  test_delay?: TestDelay | undefined;
  video_frame?: VideoFrame | undefined;
  login_request?: LoginRequest | undefined;
  login_response?: LoginResponse | undefined;
  hash?: Hash | undefined;
  mouse_event?: MouseEvent | undefined;
  audio_frame?: AudioFrame | undefined;
  cursor_data?: CursorData | undefined;
  cursor_position?: CursorPosition | undefined;
  cursor_id?: number | undefined;
  key_event?: KeyEvent | undefined;
  clipboard?: Clipboard | undefined;
  file_action?: FileAction | undefined;
  file_response?: FileResponse | undefined;
  misc?: Misc | undefined;
  cliprdr?: Cliprdr | undefined;
  message_box?: MessageBox | undefined;
  switch_sides_response?: SwitchSidesResponse | undefined;
  voice_call_request?: VoiceCallRequest | undefined;
  voice_call_response?: VoiceCallResponse | undefined;
  peer_info?: PeerInfo | undefined;
  pointer_device_event?: PointerDeviceEvent | undefined;
  auth_2fa?: Auth2FA | undefined;
  multi_clipboards?: MultiClipboards | undefined;
}
// ...
export const MouseEvent = {
  encode(message: MouseEvent, writer: _m0.Writer = _m0.Writer.create()): _m0.Writer {
    if (message.mask !== 0) {
      writer.uint32(8).int32(message.mask);
    }
    if (message.x !== 0) {
      writer.uint32(16).sint32(message.x);
    }
    if (message.y !== 0) {
      writer.uint32(24).sint32(message.y);
    }
    writer.uint32(34).fork();
    for (const v of message.modifiers) {
      writer.int32(v);
    }
    writer.ldelim();
    return writer;
  },

  decode(input: _m0.Reader | Uint8Array, length?: number): MouseEvent {
    const reader = input instanceof _m0.Reader ? input : _m0.Reader.create(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseMouseEvent();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1:
          if (tag !== 8) {
            break;
          }

          message.mask = reader.int32();
          continue;
        case 2:
          if (tag !== 16) {
            break;
          }

          message.x = reader.sint32();
          continue;
        case 3:
          if (tag !== 24) {
            break;
          }

          message.y = reader.sint32();
          continue;
        case 4:
          if (tag === 32) {
            message.modifiers.push(reader.int32() as any);

            continue;
          }

          if (tag === 34) {
            const end2 = reader.uint32() + reader.pos;
            while (reader.pos < end2) {
              message.modifiers.push(reader.int32() as any);
            }

            continue;
          }

          break;
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skipType(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): MouseEvent {
    return {
      mask: isSet(object.mask) ? globalThis.Number(object.mask) : 0,
      x: isSet(object.x) ? globalThis.Number(object.x) : 0,
      y: isSet(object.y) ? globalThis.Number(object.y) : 0,
      modifiers: globalThis.Array.isArray(object?.modifiers)
        ? object.modifiers.map((e: any) => controlKeyFromJSON(e))
        : [],
    };
  },

  toJSON(message: MouseEvent): unknown {
    const obj: any = {};
    if (message.mask !== 0) {
      obj.mask = Math.round(message.mask);
    }
    if (message.x !== 0) {
      obj.x = Math.round(message.x);
    }
    if (message.y !== 0) {
      obj.y = Math.round(message.y);
    }
    if (message.modifiers?.length) {
      obj.modifiers = message.modifiers.map((e) => controlKeyToJSON(e));
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<MouseEvent>, I>>(base?: I): MouseEvent {
    return MouseEvent.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<MouseEvent>, I>>(object: I): MouseEvent {
    const message = createBaseMouseEvent();
    message.mask = object.mask ?? 0;
    message.x = object.x ?? 0;
    message.y = object.y ?? 0;
    message.modifiers = object.modifiers?.map((e) => e) || [];
    return message;
  },
};